requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
urllib3==2.3.0
PyJWT==2.10.1
cryptography==44.0.2
python-dotenv==1.1.0
